More auth0 code.
