+++ Hestia/Hestia/Hestia.csproj
@@ -64,6 +64,9 @@
     <AppExtensionDebugBundleId />
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="backend\Auth0ServerInteractor.cs" />
+    <Compile Include="backend\HestiaServerInteractorRemote.cs" />
+    <Compile Include="backend\IHestiaServerInteractor.cs" />
     <Compile Include="backend\exceptions\ServerInteractionException.cs" />
     <Compile Include="backend\models\ActivatorState.cs" />
     <Compile Include="backend\models\Activator.cs" />
@@ -73,7 +76,7 @@
     <Compile Include="backend\models\PluginInfo.cs" />
     <Compile Include="backend\NetworkHandler.cs" />
     <Compile Include="backend\WebServer.cs" />
-    <Compile Include="backend\ServerInteractor.cs" />
+    <Compile Include="backend\HestiaServerInteractorLocal.cs" />
     <Compile Include="backend\utils\JsonValidator.cs" />
     <Compile Include="backend\utils\PingServer.cs" />
     <Compile Include="Main.cs" />
