+++ Hestia/Hestia/Resources/strings.resx
@@ -277,6 +277,9 @@
     <value>mainToServerConnect</value>
   </data>
   <data name="auth0ServersPath" xml:space="preserve">
-    <value>servers</value>
+    <value>servers/</value>
+  </data>
+  <data name="auth0RequestPath" xml:space="preserve">
+    <value>request</value>
   </data>
 </root>
