+++ Hestia/Hestia/backend/WebServer.cs
@@ -2,12 +2,12 @@
 {
     public class WebServer
     {
-        public ServerInteractor Interactor { get; set; }
+        public IHestiaServerInteractor Interactor { get; set; }
         public bool Selected { set; get; }
         public string Name { get; set; }
         public string Id { get; set; }
 
-        public WebServer(bool selected, ServerInteractor interactor)
+        public WebServer(bool selected, IHestiaServerInteractor interactor)
         {
             Interactor = interactor;
             Selected = selected;
